Correct the assert() order in test_libvirt_blockinfo
I'm not a big believer in changing the order of assert arguments
for the fun of it. On the other hand, this file was inconsistent
(and mostly backwards). I just wasted a while trying to figure out
what was wrong with my code to discover it was the arguments being
passed in the wrong order. So I am fixing that.

Change-Id: If5cb6c12a26fedbcd1f15c2a3f310e7c8347b586